[xlinq] add XmlSerializer for mono support.
Even after we imported System.Xml referencesource, XmlSerializer is still
based on existing code. And we have some special trick for XElement
serialization support in System.Xml, by allowing access from
System.Xml.Linq.dll to use special type converter attribute.

It does not exist in referencesource and thus we lost serializability on
XElement, so we need to add the missing code to bring it back.